SIH/Mapping/requirements.txt
rank_bm25
scikit-learn
sentence-transformers
numpy
tqdm
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
authlib
itsdangerous
httpx